backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdfium2
google-generativeai
python-dotenv
pillow